+++ Sources/iOS/NavigationBar.swift
@@ -36,6 +36,12 @@ open class NavigationBar: UINavigationBar {
         return 0 < bounds.width && 0 < bounds.height && nil != superview
     }
     
+    /// Detail UILabel when in landscape for iOS 11.
+    fileprivate var toolbarToText: [Toolbar: String?]?
+    
+    /// PulseAnimation for different views.
+    fileprivate var viewToPulseAnimation: [UIView: PulseAnimation]?
+    
     open override var intrinsicContentSize: CGSize {
         return CGSize(width: bounds.width, height: bounds.height)
     }
@@ -154,9 +160,14 @@ open class NavigationBar: UINavigationBar {
         isTranslucent = false
         depthPreset = .depth1
         contentScaleFactor = Screen.scale
-		backButtonImage = Icon.cm.arrowBack
         contentEdgeInsetsPreset = .square1
         interimSpacePreset = .interimSpace3
+        backButtonImage = Icon.cm.arrowBack
+        
+        if #available(iOS 11, *) {
+            toolbarToText = [:]
+            viewToPulseAnimation = [:]
+        }
         
         let image = UIImage()
         shadowImage = image
@@ -175,35 +186,51 @@ internal extension NavigationBar {
             return
         }
         
-        item.toolbar.backgroundColor = .clear
-        item.titleView = item.toolbar
+        let toolbar = item.toolbar
+        toolbar.backgroundColor = .clear
+        toolbar.interimSpace = interimSpace
+        toolbar.contentEdgeInsets = contentEdgeInsets
         
-        guard let v = item.titleView as? Toolbar else {
-            return
-        }
-        
-        if #available(iOS 11.0, *) {
-            let h = CGFloat(heightPreset.rawValue)
-            frame = CGRect(x: frame.origin.x, y: 20, width: frame.size.width, height: h)
+        if #available(iOS 11, *) {
+            if Application.shouldStatusBarBeHidden {
+                toolbar.contentEdgeInsetsPreset = .none
             
-            for subview in subviews {
-                var stringFromClass = NSStringFromClass(subview.classForCoder)
-                if stringFromClass.contains("BarBackground") {
-                    subview.frame = CGRect(x: 0, y: 0, width: frame.width, height: h)
+                for v in toolbar.leftViews + toolbar.rightViews {
+                    guard var b = v as? Pulseable else {
+                        continue
+                    }
+                    
+                    guard .none != b.pulseAnimation else {
+                        continue
+                    }
+                    
+                    viewToPulseAnimation?[v] = b.pulseAnimation
+                    b.pulseAnimation = .none
                 }
                 
-                stringFromClass = NSStringFromClass(subview.classForCoder)
-                if stringFromClass.contains("BarContent") {
-                    subview.frame = CGRect(x: subview.frame.origin.x, y: 0, width: subview.frame.width, height: h)
+                if nil != toolbar.detailLabel.text {
+                    toolbarToText?[toolbar] = toolbar.detailLabel.text
+                    toolbar.detailLabel.text = nil
+                }
+            } else if nil != toolbarToText?[toolbar] {
+                toolbar.detailLabel.text = toolbarToText?[toolbar] ?? nil
+                toolbarToText?[toolbar] = nil
+                
+                for v in toolbar.leftViews + toolbar.rightViews {
+                    guard var b = v as? Pulseable else {
+                        continue
+                    }
+                    
+                    guard let a = viewToPulseAnimation?[v] else {
+                        continue
+                    }
+                    
+                    b.pulseAnimation = a
                 }
             }
-            
-            v.frame = frame
-        } else {
-            v.frame = bounds
         }
         
-        v.contentEdgeInsets = contentEdgeInsets
-        v.interimSpace = interimSpace
+        item.titleView = toolbar
+        item.titleView!.frame = frame
     }
 }